Fix interface to PanResponder.t so it is possible to use BsReasonReact.PanResponder.panHandlers
