--- lib/js/src/private/PanResponderUtils.js
@@ -1,23 +1,0 @@
-// Generated by BUCKLESCRIPT VERSION 2.1.0, PLEASE EDIT WITH CARE
-'use strict';
-
-var Js_primitive = require("bs-platform/lib/js/js_primitive.js");
-
-function panHandlers(t) {
-  var jsHandlers = t.panHandlers;
-  return /* record */[
-          /* onMoveShouldSetResponder */Js_primitive.undefined_to_opt(jsHandlers.onMoveShouldSetResponder),
-          /* onMoveShouldSetResponderCapture */Js_primitive.undefined_to_opt(jsHandlers.onMoveShouldSetResponderCapture),
-          /* onResponderGrant */Js_primitive.undefined_to_opt(jsHandlers.onResponderGrant),
-          /* onResponderMove */Js_primitive.undefined_to_opt(jsHandlers.onResponderMove),
-          /* onResponderReject */Js_primitive.undefined_to_opt(jsHandlers.onResponderReject),
-          /* onResponderRelease */Js_primitive.undefined_to_opt(jsHandlers.onResponderRelease),
-          /* onResponderTerminate */Js_primitive.undefined_to_opt(jsHandlers.onResponderTerminate),
-          /* onResponderTerminationRequest */Js_primitive.undefined_to_opt(jsHandlers.onResponderTerminationRequest),
-          /* onStartShouldSetResponder */Js_primitive.undefined_to_opt(jsHandlers.onStartShouldSetResponder),
-          /* onStartShouldSetResponderCapture */Js_primitive.undefined_to_opt(jsHandlers.onStartShouldSetResponderCapture)
-        ];
-}
-
-exports.panHandlers = panHandlers;
-/* No side effect */
