--- lib/js/src/private/panResponderRe.js
@@ -1,81 +1,0 @@
-// Generated by BUCKLESCRIPT VERSION 2.1.0, PLEASE EDIT WITH CARE
-'use strict';
-
-var List                     = require("bs-platform/lib/js/list.js");
-var Curry                    = require("bs-platform/lib/js/curry.js");
-var Js_undefined             = require("bs-platform/lib/js/js_undefined.js");
-var ReactNative              = require("react-native");
-var AnimatedRe$BsReactNative = require("../animatedRe.js");
-
-function callback(x, $$event, state) {
-  return Curry._2(x, $$event.nativeEvent, /* record */[
-              /* stateID */state.stateID,
-              /* moveX */state.moveX,
-              /* moveY */state.moveY,
-              /* x0 */state.x0,
-              /* y0 */state.y0,
-              /* dx */state.dx,
-              /* dy */state.dy,
-              /* vx */state.vx,
-              /* vy */state.vy,
-              /* numberActiveTouches */state.numberActiveTouches
-            ]);
-}
-
-function animatedEvent(l) {
-  var config = List.fold_left((function (x, y) {
-          var variant = y[0];
-          if (variant !== 89) {
-            if (variant >= 19713) {
-              var value = y[1];
-              return Object.assign(x, {
-                          dx: AnimatedRe$BsReactNative.ValueXY[/* getX */15](value),
-                          dy: AnimatedRe$BsReactNative.ValueXY[/* getY */16](value)
-                        });
-            } else {
-              return Object.assign(x, {
-                          dx: y[1]
-                        });
-            }
-          } else {
-            return Object.assign(x, {
-                        dy: y[1]
-                      });
-          }
-        }), { }, l);
-  return AnimatedRe$BsReactNative.$$event(/* array */[
-              null,
-              config
-            ], {
-              useNativeDriver: false
-            });
-}
-
-function create(onMoveShouldSetPanResponder, onMoveShouldSetPanResponderCapture, onStartShouldSetPanResponder, onStartShouldSetPanResponderCapture, onPanResponderReject, onPanResponderGrant, onPanResponderStart, onPanResponderEnd, onPanResponderRelease, onPanResponderMove, onPanResponderTerminate, onPanResponderTerminationRequest, onShouldBlockNativeResponder, _) {
-  var tmp;
-  if (onPanResponderMove) {
-    var x = onPanResponderMove[0];
-    tmp = x[0] >= 117802505 ? animatedEvent(x[1]) : x[1];
-  } else {
-    tmp = undefined;
-  }
-  return ReactNative.PanResponder.create({
-              onMoveShouldSetPanResponder: Js_undefined.from_opt(onMoveShouldSetPanResponder),
-              onMoveShouldSetPanResponderCapture: Js_undefined.from_opt(onMoveShouldSetPanResponderCapture),
-              onStartShouldSetPanResponder: Js_undefined.from_opt(onStartShouldSetPanResponder),
-              onStartShouldSetPanResponderCapture: Js_undefined.from_opt(onStartShouldSetPanResponderCapture),
-              onPanResponderReject: Js_undefined.from_opt(onPanResponderReject),
-              onPanResponderGrant: Js_undefined.from_opt(onPanResponderGrant),
-              onPanResponderStart: Js_undefined.from_opt(onPanResponderStart),
-              onPanResponderEnd: Js_undefined.from_opt(onPanResponderEnd),
-              onPanResponderRelease: Js_undefined.from_opt(onPanResponderRelease),
-              onPanResponderMove: tmp,
-              onPanResponderTerminate: Js_undefined.from_opt(onPanResponderTerminate),
-              onPanResponderTerminationRequest: Js_undefined.from_opt(onPanResponderTerminationRequest),
-              onShouldBlockNativeResponder: Js_undefined.from_opt(onShouldBlockNativeResponder)
-            });
-}
-
-exports.callback = callback;
-exports.create   = create;
-/* react-native Not a pure module */
