+++ src/panResponder.re
@@ -1,3 +1,132 @@
-include PanResponderRe;
+type t;
 
-include PanResponderUtils;
+type gestureState = {
+  stateID: float,
+  moveX: float,
+  moveY: float,
+  x0: float,
+  y0: float,
+  dx: float,
+  dy: float,
+  vx: float,
+  vy: float,
+  numberActiveTouches: int
+};
+
+[@bs.scope "PanResponder"] [@bs.module "react-native"] external _create : 'a => t = "create";
+
+type jsGestureState = {
+  .
+  "stateID": float,
+  "moveX": float,
+  "moveY": float,
+  "x0": float,
+  "y0": float,
+  "dx": float,
+  "dy": float,
+  "vx": float,
+  "vy": float,
+  "numberActiveTouches": int
+};
+
+type callback('a) = ({. "nativeEvent": RNEvent.NativeEvent.t}, jsGestureState) => 'a;
+
+let callback = (x: (RNEvent.NativeEvent.t, gestureState) => 'a) : callback('a) =>
+  (event, state) =>
+    x(
+      event##nativeEvent,
+      {
+        dx: state##dx,
+        dy: state##dy,
+        moveX: state##moveX,
+        moveY: state##moveY,
+        numberActiveTouches: state##numberActiveTouches,
+        stateID: state##stateID,
+        vx: state##vx,
+        vy: state##vy,
+        x0: state##x0,
+        y0: state##y0
+      }
+    );
+
+external shamelesslyWrapCallback : 'a => callback(unit) = "%identity";
+
+let animatedEvent = (l) => {
+  let config =
+    List.fold_left(
+      (x, y) =>
+        switch y {
+        | `X((value: AnimatedRe.Value.t)) => Js.Obj.assign(x, {"dx": value})
+        | `Y((value: AnimatedRe.Value.t)) => Js.Obj.assign(x, {"dy": value})
+        | `XY(value) =>
+          Js.Obj.assign(x, AnimatedRe.ValueXY.({"dx": getX(value), "dy": getY(value)}))
+        },
+      Js.Obj.empty(),
+      l
+    );
+  shamelesslyWrapCallback(
+    AnimatedRe.event([|Js.null, Js.Null.return(config)|], {"useNativeDriver": Js.false_})
+  )
+};
+
+let create =
+    (
+      ~onMoveShouldSetPanResponder: option(callback(bool))=?,
+      ~onMoveShouldSetPanResponderCapture: option(callback(bool))=?,
+      ~onStartShouldSetPanResponder: option(callback(bool))=?,
+      ~onStartShouldSetPanResponderCapture: option(callback(bool))=?,
+      ~onPanResponderReject: option(callback(unit))=?,
+      ~onPanResponderGrant: option(callback(unit))=?,
+      ~onPanResponderStart: option(callback(unit))=?,
+      ~onPanResponderEnd: option(callback(unit))=?,
+      ~onPanResponderRelease: option(callback(unit))=?,
+      ~onPanResponderMove=?,
+      ~onPanResponderTerminate: option(callback(unit))=?,
+      ~onPanResponderTerminationRequest: option(callback(bool))=?,
+      ~onShouldBlockNativeResponder: option(callback(bool))=?,
+      ()
+    ) =>
+  _create({
+    "onMoveShouldSetPanResponder": Js.Undefined.from_opt(onMoveShouldSetPanResponder),
+    "onMoveShouldSetPanResponderCapture": Js.Undefined.from_opt(onMoveShouldSetPanResponderCapture),
+    "onStartShouldSetPanResponder": Js.Undefined.from_opt(onStartShouldSetPanResponder),
+    "onStartShouldSetPanResponderCapture":
+      Js.Undefined.from_opt(onStartShouldSetPanResponderCapture),
+    "onPanResponderReject": Js.Undefined.from_opt(onPanResponderReject),
+    "onPanResponderGrant": Js.Undefined.from_opt(onPanResponderGrant),
+    "onPanResponderStart": Js.Undefined.from_opt(onPanResponderStart),
+    "onPanResponderEnd": Js.Undefined.from_opt(onPanResponderEnd),
+    "onPanResponderRelease": Js.Undefined.from_opt(onPanResponderRelease),
+    "onPanResponderMove":
+      switch onPanResponderMove {
+      | None => Js.undefined
+      | Some(x) =>
+        switch x {
+        | `update(l) => Js.Undefined.return(animatedEvent(l))
+        | `callback((x: callback(unit))) => Js.Undefined.return(x)
+        }
+      },
+    "onPanResponderTerminate": Js.Undefined.from_opt(onPanResponderTerminate),
+    "onPanResponderTerminationRequest": Js.Undefined.from_opt(onPanResponderTerminationRequest),
+    "onShouldBlockNativeResponder": Js.Undefined.from_opt(onShouldBlockNativeResponder)
+  });
+
+[@bs.get] external _panHandlers : t => Js.t('a) = "panHandlers";
+
+let panHandlers = (t) : Props.touchResponderHandlers => {
+  let jsHandlers = _panHandlers(t);
+  {
+    onMoveShouldSetResponder: Js.Undefined.to_opt(jsHandlers##onMoveShouldSetResponder),
+    onMoveShouldSetResponderCapture:
+      Js.Undefined.to_opt(jsHandlers##onMoveShouldSetResponderCapture),
+    onResponderGrant: Js.Undefined.to_opt(jsHandlers##onResponderGrant),
+    onResponderMove: Js.Undefined.to_opt(jsHandlers##onResponderMove),
+    onResponderReject: Js.Undefined.to_opt(jsHandlers##onResponderReject),
+    onResponderRelease: Js.Undefined.to_opt(jsHandlers##onResponderRelease),
+    onResponderTerminate: Js.Undefined.to_opt(jsHandlers##onResponderTerminate),
+    onResponderTerminationRequest: Js.Undefined.to_opt(jsHandlers##onResponderTerminationRequest),
+    onStartShouldSetResponder: Js.Undefined.to_opt(jsHandlers##onStartShouldSetResponder),
+    onStartShouldSetResponderCapture:
+      Js.Undefined.to_opt(jsHandlers##onStartShouldSetResponderCapture)
+  }
+};
