--- src/private/PanResponderUtils.re
@@ -1,19 +1,0 @@
-[@bs.get] external _panHandlers : PanResponderRe.t => Js.t('a) = "panHandlers";
-
-let panHandlers = (t) : Props.touchResponderHandlers => {
-  let jsHandlers = _panHandlers(t);
-  {
-    onMoveShouldSetResponder: Js.Undefined.to_opt(jsHandlers##onMoveShouldSetResponder),
-    onMoveShouldSetResponderCapture:
-      Js.Undefined.to_opt(jsHandlers##onMoveShouldSetResponderCapture),
-    onResponderGrant: Js.Undefined.to_opt(jsHandlers##onResponderGrant),
-    onResponderMove: Js.Undefined.to_opt(jsHandlers##onResponderMove),
-    onResponderReject: Js.Undefined.to_opt(jsHandlers##onResponderReject),
-    onResponderRelease: Js.Undefined.to_opt(jsHandlers##onResponderRelease),
-    onResponderTerminate: Js.Undefined.to_opt(jsHandlers##onResponderTerminate),
-    onResponderTerminationRequest: Js.Undefined.to_opt(jsHandlers##onResponderTerminationRequest),
-    onStartShouldSetResponder: Js.Undefined.to_opt(jsHandlers##onStartShouldSetResponder),
-    onStartShouldSetResponderCapture:
-      Js.Undefined.to_opt(jsHandlers##onStartShouldSetResponderCapture)
-  }
-};
