--- src/private/PanResponderUtils.rei
@@ -1,1 +1,0 @@
-let panHandlers: PanResponderRe.t => Props.touchResponderHandlers;
